travelAgent/requirements.txt
python-dotenv
requests
langchain_core
langchain
transformers
torch
langchain_community
langgraph